Week2/requirements.txt
# LangChain RAG Sample Application - Requirements
# ===============================================
# Install with: pip install -r requirements.txt

# Core LangChain packages (latest versions)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0

# For agent functionality
langgraph>=0.2.0

# Vector store
chromadb>=0.5.0

# Document loaders
beautifulsoup4>=4.12.0
pypdf>=4.0.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests (for web loading)
requests>=2.31.0

# Optional: For local LLM support
# langchain-ollama>=0.2.0

# Optional: Alternative vector stores
# faiss-cpu>=1.8.0
# pinecone-client>=3.0.0

# Optional: Additional document loaders
# unstructured>=0.15.0
# python-docx>=1.1.0
